Reorganising package structure
