+++ cagecleaner/classes.py
@@ -6,12 +6,9 @@ from cagecleaner import util
 
 # External libraries:
 import pandas as pd
-import numpy as np
 import os
-import sys
 import subprocess
 import gzip
-import re
 import shutil
 from tempfile import TemporaryDirectory
 from abc import ABC, abstractmethod
@@ -19,7 +16,6 @@ from pathlib import Path
 from copy import deepcopy
 from scipy.stats import zscore
 from random import choice
-from itertools import batched
 from Bio import SeqIO
 from cblaster.classes import Session
 from importlib import resources
@@ -124,27 +120,6 @@ Run --|                                                                         
         self.DEREP_OUT_DIR.mkdir(parents = True, exist_ok = True)
         self.GENOME_DIR.mkdir(parents = True, exist_ok = True)
 
-    @staticmethod
-    def fromArgs(args):
-        """
-        This is the first entry point for initializing a run.
-        We read the session file, determine its mode, and return a corresponding LocalRun or RemoteRun
-        """
-        print("\n--- Loading session file. ---")
-        mode = Session.from_file(args.session_file).params['mode']
-        
-        if mode == 'local' or mode == 'hmm':
-            print(f"Detected {mode} mode.")
-            return LocalRun(args)
-        
-        elif mode == 'remote':
-            print("Detected remote mode.")
-            return RemoteRun(args)
-        
-        else:
-            print(f"CAGEcleaner does not support cblaster {mode} mode for the moment. Exiting the program.")
-            sys.exit()
-    
     def dereplicate(self):
         """
         This method is the entry point for the dereplication.
@@ -461,438 +436,3 @@ Run --|                                                                         
         else:
             pass
     
- #####################################################################################################
-   
-class LocalRun(Run):
-    
-    def __init__(self, args):
-                
-        # Call the parent class initiator
-        super().__init__(args)
-        
-        # Can't keep downloads in local mode:
-        assert self.keep_downloads == False, "Can't keep downloads in local mode."
-        # Also make sure keep_intermediate doesn't fail:
-        if self.keep_intermediate:
-            self.keep_intermediate = False
-            self.keep_dereplication = True
-        
-        # Make sure there is no exotic stuff in the provided genome folder
-        for file in self.USER_GENOME_DIR.iterdir():
-            assert file.is_file(), f"The following object in the genome folder is not a file: {file}. Please move or remove it."
-            assert (util.isFasta(str(file)) or util.isGff(str(file)) or util.isGenbank(str(file))), f"The following file is not in the correct format: {file}.\nWe only accept the following suffices: [.fasta, .fna, .fa, .gff3, .gff, .gbff, .gbk, .gb]. The '.gz' extension is allowed."
-        
-        # Remove organisms that the user wants to be excluded:
-        if self.excluded_organisms != {''}:
-            self.VERBOSE(f"Excluding the following organisms: {', '.join(self.excluded_organisms)}")
-            self.binary_df = self.binary_df[~self.binary_df['Organism'].isin(self.excluded_organisms)]
-
-        # Remove scaffold IDs specified by the user:
-        if self.excluded_scaffolds != {''}:
-            self.VERBOSE(f"Excluding the following scaffolds: {', '.join(self.excluded_scaffolds)}")
-            # Here the approach slightly differs as users might have provided prefixed scaffold IDs:
-            # Add a column with a prefixed scaffold based on the Organism column
-            self.binary_df['prefixed_scaffold'] = self.binary_df['Organism'] + ':' + self.binary_df['Scaffold']
-            # If the prefixed scaffold ends with any of the strings in the set of scaffolds to exclude, remove it:
-            self.binary_df = self.binary_df[self.binary_df['prefixed_scaffold'].str.endswith(tuple(self.excluded_scaffolds)) == False]
-            # Clean up:
-            self.binary_df = self.binary_df.drop(columns=['prefixed_scaffold'])
-            
-    def prepareGenomes(self) -> None:
-        """
-        This function is called to inspect the provided genome folder and convert GenBank files to FASTA files if necessary.
-        If a FASTA file is found, it is assumed that all genomes are stored in the provided folder in FASTA format.
-        If no FASTA is found, but instead a GenBank is found, it is assumed that all genomes are in GenBank format.
-        Temporary FASTA-converted copies of these GenBank files will be placed in the temporary folders.
-        Only FASTA files can be passed on to skDER.
-        If neither GenBank nor FASTA is found, the program exits.
-    
-        Mutates:
-            self.GENOME_DIR: Path: Path to the genome directory
-        """
-        
-        # Assert that Organism and filenames correspond:
-        files_in_genomes_dir: set = {util.removeSuffixes(file) for file in os.listdir(self.USER_GENOME_DIR)}
-        organisms_in_session: set = {util.removeSuffixes(organism.name) for organism in self.session.organisms}
-        
-        assert files_in_genomes_dir >= organisms_in_session, "The genomes of all organisms in the cblaster session have not been found in the genome directory. Check the paths and please make sure you have not changed the genome filenames between a cblaster run and a CAGEcleaner run."
-        
-        # Check if there are FASTA files in the genome folder:
-        fasta_in_folder = [util.isFasta(str(file)) for file in self.USER_GENOME_DIR.iterdir()]
-        genbank_in_folder = [util.isGenbank(str(file)) for file in self.USER_GENOME_DIR.iterdir()]
-
-        if any(fasta_in_folder):
-            # In this case the genome folder path should remain the same
-            print(f"Detected {sum(fasta_in_folder)} FASTA files in {self.USER_GENOME_DIR}. These will be used for dereplication.")
-            # Redirect the genome dir to the user-provided folder:
-            self.GENOME_DIR = self.USER_GENOME_DIR
-            
-        
-        elif any(genbank_in_folder):
-            # In this case we convert to FASTA and redirect to genome folder, which is in the temp folder by default.
-            print(f"Detected {sum(genbank_in_folder)} GenBank files in {self.USER_GENOME_DIR}. Now converting to FASTA for dereplication.")
-            # Convert to FASTA files:
-            self.GENOME_DIR.mkdir(exist_ok=True)  # Make the output folder if it does not exist already.
-            util.convertGenbankToFasta(self.USER_GENOME_DIR, self.GENOME_DIR)
-            print(f"Migrated genomes in FASTA format to {self.GENOME_DIR}")
-            
-        else:
-            # If there are no FASTA or GenBank files, the program cannot proceed:
-            print("No FASTA files or GenBank files were detected in the provided genome folder. Exiting the program.")
-            sys.exit()
-            
-        assembly_files = [acc + ".fasta.gz" for acc in self.binary_df['Organism']]
-        self.binary_df['assembly_file'] = assembly_files
-        
-        return None
-    
-    def mapDereplicationToBinary(self) -> None:
-        """
-        After dereplication, map the dereplication clustering table to the binary table.
-        The dereplication clustering table is converted to a dataframe and joined with the binary table based on
-        assembly ID (full genome dereplication) or scaffold ID (region dereplication).  
-        
-        Mutates:
-            self.binary_df: pd.DataFrame: The binary table derived from a cblaster Session object.
-        """
-        # Full genome dereplication using skDER
-        if not(self.regions):
-            def extractAssembly(file_path: str) -> str:
-                return util.removeSuffixes(os.path.basename(file_path))
-            
-            def renameLabel(label: str) -> str:
-                mapping = {'representative_to_self': 'dereplication_representative',
-                           'within_cutoffs_requested': 'redundant'}
-                return mapping[label]
-            
-            self.VERBOSE("Reading skDER clustering table.")
-            # Read the skder out clustering table:
-            path_to_cluster_file: Path = self.DEREP_OUT_DIR / 'skDER_Clustering.txt'
-            # Convert to dataframe:
-            derep_df: pd.DataFrame = pd.read_table(path_to_cluster_file,
-                                     converters = {'assembly': extractAssembly,
-                                                   'representative': extractAssembly,
-                                                   'dereplication_status': renameLabel},
-                                     names = ['assembly', 'representative', 'dereplication_status'],
-                                     usecols = [0,1,4], header = 0, index_col = 'assembly'
-                                     )
-            # Join with binary df on Organism column. 
-            # Every Organism row is retained (left join).
-            # If there is a match between binary_df['Organism'] and derep_df['assembly'] (index), the representative and status is added.
-            self.VERBOSE("Joining skDER clustering table and cblaster binary table.")
-            self.binary_df = self.binary_df.join(derep_df, on='Organism')
-        
-        # Region dereplication using MMseqs2
-        else:
-            # Read the MMseqs clustering table
-            path_to_cluster_file: Path = self.TEMP_DIR / "derep_out_cluster.tsv"
-            # Convert to a dataframe
-            derep_df: pd.DataFrame = pd.read_table(path_to_cluster_file,
-                                     names = ['representative', 'Scaffold'],
-                                     header = None, index_col = 'Scaffold'
-                                     )
-            # Add dereplication status column based on whether a scaffold's ID is the same as the representative's one
-            # Every assembly_file row is retained (left join).
-            # If there is a match between binary_df['assembly_file'] and derep_df['assembly'] (its index column), the representative and status is added.
-            derep_df['dereplication_status'] = derep_df.index == derep_df['representative']
-            derep_df['dereplication_status'] = np.where(derep_df['dereplication_status'], 'dereplication_representative', 'redundant')
-            self.VERBOSE("Joining MMseqs2 clustering table and cblaster binary table.")
-            self.binary_df = self.binary_df.join(derep_df, on = "Scaffold")
-            
-        self.binary_df = self.binary_df.sort_values(['representative', 'dereplication_status'])           
-        print("Mapping done!")
-        
-        return None
-    
-    def run(self):
-        """
-        Run the entire LocalRun workflow.
-        """
-        print("\n--- STEP 1: Staging genomes for dereplication. ---")
-        self.prepareGenomes()
-        
-        print("\n--- STEP 2: Dereplicating. ---")
-        self.dereplicate()
-        
-        print("\n--- STEP 3: Mapping dereplication output to binary table. ---")
-        self.mapDereplicationToBinary()
-        
-        print("\n--- STEP 4: Recovering hit diversity. ---")
-        self.recoverHits()
-        
-        print("\n--- STEP 5: Filtering session file. ---")
-        self.filterSession()
-        
-        print("\n--- STEP 6: Generating output files")
-        self.generateOutput()
-        
-        # Remove the temporary directory:
-        print("Cleaning up temporary directory.")
-        self.TEMP_DIR_CONTEXT.cleanup()
-
- #####################################################################################################
-
-class RemoteRun(Run):
-    
-    def __init__(self, args):
-        # Call the parent constructor:
-        super().__init__(args)
-        
-        # Defensive check:
-        assert args.download_batch > 0, "Download batch should be larger than 0."
-        
-        # Set path to accessions and download script:
-        self.ACCESSIONS_SCRIPT: Path = Path(resources.files(__name__)) / 'get_accessions.sh'  # Path to the script that maps scaffold IDs to assembly IDs
-        self.DOWNLOAD_SCRIPT: Path = Path(resources.files(__name__)) / 'download_assemblies.sh'  # Path to the script that download the genomes from given assembly IDs
-        
-        # Variable to store assembly accessions:
-        self.assembly_accessions: list = []
-        
-        # Dictionary to store the scaffold:assembly mapping:
-        self.scaffold_assembly_pairs: dict = {}
-        
-        # Remove Organisms specified by the user:
-        if self.excluded_organisms != {''}:
-            # Replace colons with spaces to get the correct matching in the binary table.
-            self.excluded_organisms = {org.replace(':', ' ', regex=False) for org in self.excluded_organisms}
-            self.VERBOSE(f"Excluding the following organisms: {', '.join(self.excluded_organisms)}")
-            # Exclude them:
-            self.binary_df = self.binary_df[~self.binary_df['Organism'].isin(self.excluded_organisms)]
-        
-        # Remove scaffolds that the user wants excluded:
-        if self.excluded_scaffolds != {''}:
-            self.VERBOSE(f"Excluding the following scaffolds: {', '.join(self.excluded_scaffolds)}")
-            self.binary_df = self.binary_df[~self.binary_df['Scaffold'].isin(self.excluded_scaffolds)]
-        
-        # Replace colons in the bypass assemblies as wel:
-        if self.bypass_organisms != {''}:
-            self.bypass_organisms = {org.replace(':', ' ') for org in self.bypass_organisms}
-        
-    def fetchAssemblyIDs(self) -> None:
-        """
-        This function writes the scaffold IDs from the binary table to a file.
-        It then calls a bash script that fetches the assembly ID for each scaffold ID using NCBI entrez-direct utilities.
-        The results are then written to a file by the bash script, read by this python file, and stored internally as a list of assembly IDs.
-        
-        Mutates:
-            self.assembly_accessions: list: A list of assembly IDs to be downloaded later on.
-        """
-        # First we extract the scaffold IDs out of the binary table:
-        scaffolds = self.binary_df['Scaffold'].to_list()
-        
-        # Write these to a file with every scaffold ID on a new line:
-        scaffolds_file: Path = self.TEMP_DIR / 'scaffolds.txt'
-        with scaffolds_file.open('w') as file:
-            file.writelines('\n'.join(scaffolds))
-
-        # Now we use the helper bash script to map the scaffolds to the NCBI assembly IDs that host them.
-        home = os.getcwd()
-        os.chdir(self.TEMP_DIR)
-        subprocess.run(['bash', str(self.ACCESSIONS_SCRIPT), 'scaffolds.txt'], check = True)
-        os.chdir(home)
-        
-        # Read the result file
-        result_file: Path = self.TEMP_DIR / 'assembly_accessions.txt'
-        with result_file.open('r') as file:
-            # Store the assembly accessions internally:
-            self.assembly_accessions = [line.rstrip() for line in file.readlines()]
-
-        return None
-    
-    def downloadGenomes(self) -> None:
-        """
-        This function writes the assembly IDs found by fetchAssemblyIDs to a file in batches (default size of 300). Each line in the file is a batch.
-        It then calls a bash script that downloads the genomes for each line of assembly IDs.
-        Genome files are placed in the TEMP_DIR/genomes folder (in gzipped format)
-        """
-        # First we cut off the version digits of our assembly IDs, and rely on NCBI datatsets to fetch the latest version:
-        versionless_assemblies = [acc.split('.')[0] for acc in self.assembly_accessions]
-        
-        self.VERBOSE("Writing assembly ID batches for download script.")
-        # Now we write the assembly accesions to a file in batches:
-        download_batches_file: Path = self.TEMP_DIR / 'download_batches.txt'
-        with download_batches_file.open('w') as file:
-            download_batches = list(batched(versionless_assemblies, self.download_batch))
-            for batch in download_batches:
-                file.write(' '.join(batch) + '\n')
-                
-        # Run the bash script to download genomes:
-        home = os.getcwd()
-        os.chdir(self.TEMP_DIR)
-        self.VERBOSE("Calling download script.")
-        subprocess.run(["bash", str(self.DOWNLOAD_SCRIPT)], check=True)
-        os.chdir(home)
-
-        return None
-    
-    def mapScaffoldsToAssemblies(self) -> None:
-        """
-        This function maps every scaffold in the binary table to its corresponding assembly file.
-        The mapping is stored internally as a dictionary.
-        It loops over every file in the genome folder and extracts the set of scaffold using BioPython SeqIO module.
-        This set is then compared to the set of scaffolds in the binary table, and the intersection of both sets provied us the correct mapping.
-        Scaffolds are stripped off their prefixes because NCBI sometimes omits these when downloading genomes.
-        
-        Mutates:
-            self.scaffold_assembly_pairs: dict: Dictionary mapping of scaffold:assembly_file pairs
-        
-        """
-        def removePrefix(scaffold: str) -> str:
-            # ^ matches beginning of sting
-            # [^_]+ matches one or more characters that are not an underscore
-            # _ matches the underscore itself
-            pattern = r'^[^_]+_' 
-            # Replace the prefix with an empty string:
-            return re.sub(pattern, '', scaffold)
-        
-        def addPrefix(deprefixed_scaffold: str, scaffolds_in_host_assembly: set) -> str:
-            # Map back to the prefixed scaffold by scanning the assembly that hosts the given scaffold.
-            return [s for s in scaffolds_in_host_assembly if deprefixed_scaffold in s][0]
-
-        # Define the set of scaffolds (without prefix) that can be found in the binary:
-        scaffolds_in_binary: set = {scaffold.strip() for scaffold in self.binary_df['Scaffold'].to_list()}
-        # Scaffold set in binary without any prefixes:
-        scaffolds_in_binary_no_prefix: set = {removePrefix(scaffold) for scaffold in scaffolds_in_binary}
-        
-        # Loop over the directory containing all genomes:
-        for file in self.GENOME_DIR.iterdir():
-            # Only read fasta files:
-            if util.isFasta(file.name):
-                self.VERBOSE(f"Reading {file.name}")
-                # Open the file:
-                with gzip.open(file, 'rt') as assembly:
-                    # Extract the set of scaffold IDs in the file:
-                    scaffolds_in_this_assembly: set = {record.id.strip() for record in SeqIO.parse(assembly, 'fasta')}
-                    # Remove the prefixes:
-                    scaffolds_in_this_assembly_no_prefix: set = {removePrefix(scaffold) for scaffold in scaffolds_in_this_assembly}
-                    # Now we take the intersection of both sets. All the scaffolds in this intersection can be mapped to the current file in the loop:
-                    found_scaffolds_no_prefix: set = scaffolds_in_binary_no_prefix.intersection(scaffolds_in_this_assembly_no_prefix)
-                    # Now we have to add the prefix again by using the original scaffold list from the host assembly:
-                    found_scaffolds: set = {addPrefix(scaffold, scaffolds_in_this_assembly) for scaffold in found_scaffolds_no_prefix}
-                    # Tell the user what we found in this assembly file
-                    self.VERBOSE(f"Found {','.join(found_scaffolds)}") if len(found_scaffolds) > 0 else self.VERBOSE("No scaffolds found.")
-                    # Finalize the mapping in a dictionary:
-                    for scaffold in found_scaffolds:
-                        self.scaffold_assembly_pairs[scaffold] = file.name
-                        
-        return None
-                
-    def mapAssembliesToBinary(self) -> None:
-        """
-        This function maps each row in the binary table to a corresponding assembly file based on the mapping obtained by mapScaffoldsToAssemblies().
-        
-        Mutates:
-            self.binary_df: pd.DataFrame: Internal representation of the binary table.
-        
-        """
-        # Read the dictionary mapping as a dataframe with the scaffold IDs as index and the assembly file to which it belongs as 'assembly_file':
-        scaffold_assembly_pairs_df: pd.DataFrame = pd.DataFrame.from_dict(self.scaffold_assembly_pairs, orient='index', columns = ['assembly_file'])
-        
-        # Join the binary table on the 'Scaffold' column:
-        self.binary_df = self.binary_df.join(scaffold_assembly_pairs_df, on='Scaffold')
-            
-        # Extract scaffolds that could not be linked to an assembly:
-        scaffolds_with_na = self.binary_df[self.binary_df['assembly_file'].isna()]['Scaffold'].to_list()
-        
-        if scaffolds_with_na:
-            print(f"The following {len(scaffolds_with_na)} scaffolds could not be linked to a genome assembly: {', '.join(scaffolds_with_na)}. Omitting for further analysis.") 
-            # Drop the NA values:
-            self.binary_df = self.binary_df.dropna()
-        
-        return None
-    
-    def mapDereplicationToBinary(self) -> None:
-        """
-        After dereplication, map the dereplication clustering table to the binary table.
-        The dereplication clustering table is converted to a dataframe and joined with the binary table based on
-        assembly ID (full genome dereplication) or scaffold ID (region dereplication).        
-        Mutates:
-            self.binary_df: pd.DataFrame: Internal representation of the binary table.
-        """
-        # Full genome dereplication using skDER
-        if not(self.regions):
-            def extractFileName(file_path: str) -> str:
-                # Extract basename from full file path
-                return Path(file_path).name
-            
-            def renameLabel(label: str) -> str:
-                # Rename some of the skDER labels
-                mapping = {'representative_to_self': 'dereplication_representative',
-                           'within_cutoffs_requested': 'redundant'}
-                return mapping[label]
-            
-            self.VERBOSE("Reading skDER clustering table.")
-            # Read the skder out clustering table:
-            path_to_cluster_file: Path = self.DEREP_OUT_DIR / 'skDER_Clustering.txt'
-            # Convert to dataframe:
-            derep_df: pd.DataFrame = pd.read_table(path_to_cluster_file,
-                                     converters = {'assembly': extractFileName,
-                                                   'representative': extractFileName,
-                                                   'dereplication_status': renameLabel},
-                                     names = ['assembly', 'representative', 'dereplication_status'],
-                                     usecols = [0,1,4], header = 0, index_col = 'assembly'
-                                     )
-            # Join with binary df on assembly_file column. 
-            # Every assembly_file row is retained (left join).
-            # If there is a match between binary_df['assembly_file'] and derep_df['assembly'] (its index column), the representative and status is added.
-            self.VERBOSE("Joining skDER clustering table and cblaster binary table.")
-            self.binary_df = self.binary_df.join(derep_df, on='assembly_file')
-        
-        # Region dereplication using MMseqs2
-        else:
-            # Read the MMseqs2 clustering table:
-            path_to_cluster_file: Path = self.TEMP_DIR / "derep_out_cluster.tsv"
-            # Convert to dataframe
-            derep_df: pd.DataFrame = pd.read_table(path_to_cluster_file,
-                                     names = ['representative', 'Scaffold'],
-                                     header = None, index_col = 'Scaffold'
-                                     )
-            # Add dereplication status column based on whether the scaffold's ID is the same as the representative's
-            # Every assembly_file row is retained (left join).
-            # If there is a match between binary_df['assembly_file'] and derep_dfs['assembly'] (its index column), the representative and status is added.
-            derep_df['dereplication_status'] = derep_df.index == derep_df['representative']
-            derep_df['dereplication_status'] = np.where(derep_df['dereplication_status'], 'dereplication_representative', 'redundant')
-            self.VERBOSE("Joining MMseqs2 clustering table and cblaster binary table.")
-            self.binary_df = self.binary_df.join(derep_df, on = "Scaffold")
-        
-        # Sort by representative ID and then by dereplication status
-        self.binary_df = self.binary_df.sort_values(['representative', 'dereplication_status'])
-        print("Mapping done!")
-            
-        return None
-    
-    def run(self) -> None:
-        """
-        Run the entire remote workflow.
-        """
-        print("\n--- STEP 1: Fetching assembly IDs from NCBI for each scaffold ID in the cblaster binary table. ---")
-        self.fetchAssemblyIDs()  # Stores a list of NCBI assembly IDs 
-        
-        print("\n--- STEP 2: Downloading genomes for each assembly ID. ---")
-        self.downloadGenomes()  # Downloads genome for each assembly ID
-        
-        print("\n--- STEP 3: Mapping scaffold IDs to assembly IDs ---")
-        self.mapScaffoldsToAssemblies()  # Results in a dictionary of scaffold:assembly_file pairs
-        self.mapAssembliesToBinary()  # Each row in the binary table is now mapped to its assembly file
-                
-        print("\n--- STEP 4: Dereplicating ---")
-        self.dereplicate()  # Dereplicate.
-        
-        print("\n--- STEP 5: Mapping dereplication clustering to binary table ---")
-        self.mapDereplicationToBinary()  # Map each row in the binary table with its representative
-        
-        print("\n--- STEP 6: Recovering hit diversity ---")
-        self.recoverHits()  # Recover hits by content and score depending on user input
-        
-        print("\n--- STEP 7: Filtering original session file. ---")
-        self.filterSession()  # Filter the original session file, retaining only the dereplicated hits.
-        
-        print("\n--- STEP 8: Generating output files ---")
-        self.generateOutput()  # Generate all output files.
-        
-        # Remove the temporary directory:
-        print("Cleaning up temporary directory.")
-        self.TEMP_DIR_CONTEXT.cleanup()
-        
-        return None
-
